+++ R/misc.R
@@ -162,3 +162,18 @@ add_names_to_alist <- function(what) {
   names(what) <- sapply(what, function(x) deparse(x))
   lapply(what, eval)
 }
+
+#' Append to each input a listified call to that input
+#'
+#' @noRd
+append_listified <- function(input) {
+  transformed <- lapply(input, function(x) {
+    ## if (inherits(x, "list"))
+    ##   x <- list(x) ## otherwise if `x` is `list()`, the following line will leave it unchanged
+    ## as.call(c(as.name("list"), x))
+    list(x)
+  })
+  if (any(nzchar(names(transformed))))
+    names(transformed) <- sprintf("list(%s)", names(transformed))
+  c(input, transformed)
+}