+++ NEWS.md
@@ -9,6 +9,8 @@
 - Make the `what` argument to `fuzz()` accept a list of inputs, so that
   the function can test multiple inputs in the same run.
 - Add a default list of inputs to `fuzz()`.
+- Add the `listify.what` argument to `fuzz()` so that each input in `what`
+  is also tested in its listified version.
 
 # CBTF 0.2.0 (2025-07-16)